UX: rewrite paths in errors to libs sources for easier
